Move distributions to more consistent places
ZeroInflated to mixture.py
DiracDelta to distribution.py
