+++ docs/source/api/distributions/utilities.rst
@@ -11,3 +11,4 @@ Distribution utilities
     Discrete
     Distribution
     SymbolicRandomVariable
+    DiracDelta
