+++ pymc/distributions/discrete.py
@@ -46,12 +46,10 @@ from pymc.distributions.dist_math import (
     normal_lcdf,
 )
 from pymc.distributions.distribution import Discrete
-from pymc.distributions.mixture import Mixture
 from pymc.distributions.shape_utils import rv_size_is_none
 from pymc.logprob.basic import logp
 from pymc.math import sigmoid
 from pymc.pytensorf import floatX, intX
-from pymc.vartypes import continuous_types
 
 __all__ = [
     "Binomial",
@@ -60,10 +58,6 @@ __all__ = [
     "DiscreteWeibull",
     "Poisson",
     "NegativeBinomial",
-    "DiracDelta",
-    "ZeroInflatedPoisson",
-    "ZeroInflatedBinomial",
-    "ZeroInflatedNegativeBinomial",
     "DiscreteUniform",
     "Geometric",
     "HyperGeometric",
@@ -1171,313 +1165,6 @@ class Categorical(Discrete):
             p_ <= 1,
             pt.isclose(pt.sum(p, axis=-1), 1),
             msg="0 <= p <=1, sum(p) = 1",
-        )
-
-
-class DiracDeltaRV(RandomVariable):
-    name = "diracdelta"
-    ndim_supp = 0
-    ndims_params = [0]
-    _print_name = ("DiracDelta", "\\operatorname{DiracDelta}")
-
-    def make_node(self, rng, size, dtype, c):
-        c = pt.as_tensor_variable(c)
-        return super().make_node(rng, size, c.dtype, c)
-
-    @classmethod
-    def rng_fn(cls, rng, c, size=None):
-        if size is None:
-            return c.copy()
-        return np.full(size, c)
-
-
-diracdelta = DiracDeltaRV()
-
-
-class DiracDelta(Discrete):
-    r"""
-    DiracDelta log-likelihood.
-
-    Parameters
-    ----------
-    c : tensor_like of float or int
-        Dirac Delta parameter. The dtype of `c` determines the dtype of the distribution.
-        This can affect which sampler is assigned to DiracDelta variables, or variables
-        that use DiracDelta, such as Mixtures.
-    """
-
-    rv_op = diracdelta
-
-    @classmethod
-    def dist(cls, c, *args, **kwargs):
-        c = pt.as_tensor_variable(c)
-        if c.dtype in continuous_types:
-            c = floatX(c)
-        return super().dist([c], **kwargs)
-
-    def moment(rv, size, c):
-        if not rv_size_is_none(size):
-            c = pt.full(size, c)
-        return c
-
-    def logp(value, c):
-        return pt.switch(
-            pt.eq(value, c),
-            pt.zeros_like(value),
-            -np.inf,
-        )
-
-    def logcdf(value, c):
-        return pt.switch(
-            pt.lt(value, c),
-            -np.inf,
-            0,
-        )
-
-
-def _zero_inflated_mixture(*, name, nonzero_p, nonzero_dist, **kwargs):
-    """Helper function to create a zero-inflated mixture
-
-    If name is `None`, this function returns an unregistered variable
-    """
-    nonzero_p = pt.as_tensor_variable(floatX(nonzero_p))
-    weights = pt.stack([1 - nonzero_p, nonzero_p], axis=-1)
-    comp_dists = [
-        DiracDelta.dist(0),
-        nonzero_dist,
-    ]
-    if name is not None:
-        return Mixture(name, weights, comp_dists, **kwargs)
-    else:
-        return Mixture.dist(weights, comp_dists, **kwargs)
-
-
-class ZeroInflatedPoisson:
-    R"""
-    Zero-inflated Poisson log-likelihood.
-
-    Often used to model the number of events occurring in a fixed period
-    of time when the times at which events occur are independent.
-    The pmf of this distribution is
-
-    .. math::
-
-        f(x \mid \psi, \mu) = \left\{ \begin{array}{l}
-            (1-\psi) + \psi e^{-\mu}, \text{if } x = 0 \\
-            \psi \frac{e^{-\mu}\mu^x}{x!}, \text{if } x=1,2,3,\ldots
-            \end{array} \right.
-
-    .. plot::
-        :context: close-figs
-
-        import matplotlib.pyplot as plt
-        import numpy as np
-        import scipy.stats as st
-        import arviz as az
-        plt.style.use('arviz-darkgrid')
-        x = np.arange(0, 22)
-        psis = [0.7, 0.4]
-        mus = [8, 4]
-        for psi, mu in zip(psis, mus):
-            pmf = st.poisson.pmf(x, mu)
-            pmf[0] = (1 - psi) + pmf[0]
-            pmf[1:] =  psi * pmf[1:]
-            pmf /= pmf.sum()
-            plt.plot(x, pmf, '-o', label='$\\psi$ = {}, $\\mu$ = {}'.format(psi, mu))
-        plt.xlabel('x', fontsize=12)
-        plt.ylabel('f(x)', fontsize=12)
-        plt.legend(loc=1)
-        plt.show()
-
-    ========  ==========================
-    Support   :math:`x \in \mathbb{N}_0`
-    Mean      :math:`\psi\mu`
-    Variance  :math:`\mu + \frac{1-\psi}{\psi}\mu^2`
-    ========  ==========================
-
-    Parameters
-    ----------
-    psi : tensor_like of float
-        Expected proportion of Poisson variates (0 < psi < 1)
-    mu : tensor_like of float
-        Expected number of occurrences during the given interval
-        (mu >= 0).
-    """
-
-    def __new__(cls, name, psi, mu, **kwargs):
-        return _zero_inflated_mixture(
-            name=name, nonzero_p=psi, nonzero_dist=Poisson.dist(mu=mu), **kwargs
-        )
-
-    @classmethod
-    def dist(cls, psi, mu, **kwargs):
-        return _zero_inflated_mixture(
-            name=None, nonzero_p=psi, nonzero_dist=Poisson.dist(mu=mu), **kwargs
-        )
-
-
-class ZeroInflatedBinomial:
-    R"""
-    Zero-inflated Binomial log-likelihood.
-
-    The pmf of this distribution is
-
-    .. math::
-
-        f(x \mid \psi, n, p) = \left\{ \begin{array}{l}
-            (1-\psi) + \psi (1-p)^{n}, \text{if } x = 0 \\
-            \psi {n \choose x} p^x (1-p)^{n-x}, \text{if } x=1,2,3,\ldots,n
-            \end{array} \right.
-
-    .. plot::
-        :context: close-figs
-
-        import matplotlib.pyplot as plt
-        import numpy as np
-        import scipy.stats as st
-        import arviz as az
-        plt.style.use('arviz-darkgrid')
-        x = np.arange(0, 25)
-        ns = [10, 20]
-        ps = [0.5, 0.7]
-        psis = [0.7, 0.4]
-        for n, p, psi in zip(ns, ps, psis):
-            pmf = st.binom.pmf(x, n, p)
-            pmf[0] = (1 - psi) + pmf[0]
-            pmf[1:] =  psi * pmf[1:]
-            pmf /= pmf.sum()
-            plt.plot(x, pmf, '-o', label='n = {}, p = {}, $\\psi$ = {}'.format(n, p, psi))
-        plt.xlabel('x', fontsize=12)
-        plt.ylabel('f(x)', fontsize=12)
-        plt.legend(loc=1)
-        plt.show()
-
-    ========  ==========================
-    Support   :math:`x \in \mathbb{N}_0`
-    Mean      :math:`\psi n p`
-    Variance  :math:`(1-\psi) n p [1 - p(1 - \psi n)].`
-    ========  ==========================
-
-    Parameters
-    ----------
-    psi : tensor_like of float
-        Expected proportion of Binomial variates (0 < psi < 1)
-    n : tensor_like of int
-        Number of Bernoulli trials (n >= 0).
-    p : tensor_like of float
-        Probability of success in each trial (0 < p < 1).
-
-    """
-
-    def __new__(cls, name, psi, n, p, **kwargs):
-        return _zero_inflated_mixture(
-            name=name, nonzero_p=psi, nonzero_dist=Binomial.dist(n=n, p=p), **kwargs
-        )
-
-    @classmethod
-    def dist(cls, psi, n, p, **kwargs):
-        return _zero_inflated_mixture(
-            name=None, nonzero_p=psi, nonzero_dist=Binomial.dist(n=n, p=p), **kwargs
-        )
-
-
-class ZeroInflatedNegativeBinomial:
-    R"""
-    Zero-Inflated Negative binomial log-likelihood.
-
-    The Zero-inflated version of the Negative Binomial (NB).
-    The NB distribution describes a Poisson random variable
-    whose rate parameter is gamma distributed.
-    The pmf of this distribution is
-
-    .. math::
-
-       f(x \mid \psi, \mu, \alpha) = \left\{
-         \begin{array}{l}
-           (1-\psi) + \psi \left (
-             \frac{\alpha}{\alpha+\mu}
-           \right) ^\alpha, \text{if } x = 0 \\
-           \psi \frac{\Gamma(x+\alpha)}{x! \Gamma(\alpha)} \left (
-             \frac{\alpha}{\mu+\alpha}
-           \right)^\alpha \left(
-             \frac{\mu}{\mu+\alpha}
-           \right)^x, \text{if } x=1,2,3,\ldots
-         \end{array}
-       \right.
-
-    .. plot::
-        :context: close-figs
-
-        import matplotlib.pyplot as plt
-        import numpy as np
-        import scipy.stats as st
-        from scipy import special
-        import arviz as az
-        plt.style.use('arviz-darkgrid')
-
-        def ZeroInfNegBinom(a, m, psi, x):
-            pmf = special.binom(x + a - 1, x) * (a / (m + a))**a * (m / (m + a))**x
-            pmf[0] = (1 - psi) + pmf[0]
-            pmf[1:] =  psi * pmf[1:]
-            pmf /= pmf.sum()
-            return pmf
-
-        x = np.arange(0, 25)
-        alphas = [2, 4]
-        mus = [2, 8]
-        psis = [0.7, 0.7]
-        for a, m, psi in zip(alphas, mus, psis):
-            pmf = ZeroInfNegBinom(a, m, psi, x)
-            plt.plot(x, pmf, '-o', label=r'$\alpha$ = {}, $\mu$ = {}, $\psi$ = {}'.format(a, m, psi))
-        plt.xlabel('x', fontsize=12)
-        plt.ylabel('f(x)', fontsize=12)
-        plt.legend(loc=1)
-        plt.show()
-
-    ========  ==========================
-    Support   :math:`x \in \mathbb{N}_0`
-    Mean      :math:`\psi\mu`
-    Var       :math:`\psi\mu +  \left (1 + \frac{\mu}{\alpha} + \frac{1-\psi}{\mu} \right)`
-    ========  ==========================
-
-    The zero inflated negative binomial distribution can be parametrized
-    either in terms of mu or p, and either in terms of alpha or n.
-    The link between the parametrizations is given by
-
-    .. math::
-
-        \mu &= \frac{n(1-p)}{p} \\
-        \alpha &= n
-
-    Parameters
-    ----------
-    psi : tensor_like of float
-        Expected proportion of NegativeBinomial variates (0 < psi < 1)
-    mu : tensor_like of float
-        Poisson distribution parameter (mu > 0).
-    alpha : tensor_like of float
-        Gamma distribution parameter (alpha > 0).
-    p : tensor_like of float
-        Alternative probability of success in each trial (0 < p < 1).
-    n : tensor_like of float
-        Alternative number of target success trials (n > 0)
-    """
-
-    def __new__(cls, name, psi, mu=None, alpha=None, p=None, n=None, **kwargs):
-        return _zero_inflated_mixture(
-            name=name,
-            nonzero_p=psi,
-            nonzero_dist=NegativeBinomial.dist(mu=mu, alpha=alpha, p=p, n=n),
-            **kwargs,
-        )
-
-    @classmethod
-    def dist(cls, psi, mu=None, alpha=None, p=None, n=None, **kwargs):
-        return _zero_inflated_mixture(
-            name=None,
-            nonzero_p=psi,
-            nonzero_dist=NegativeBinomial.dist(mu=mu, alpha=alpha, p=p, n=n),
-            **kwargs,
         )
